+++ changelog.md
@@ -1,6 +1,7 @@
 Versie 6.0 (Trajectum
 - Afhankelijkheden: Bootstrap geüpdatet naar 4.2.1 (#75)
 - Afhankelijkheden: jQuery geüpdatet naar 3.3.1
+- Verwijderd: delen en Facebookintegratie (#80)
 - Verwijderd: ideeënbus
 - Verwijderd: instelling voor extra bodycode (kan nu via PHP-includes)
 - Verwijderd: klassiek menu (#76)